+++ R/praatScriptIntensity.R
@@ -18,6 +18,11 @@
 #' @param interpolation If sample.points are specified, this is the interpolation to use
 #'     when getting individual values. Possible values are 'nearest', 'linear', 'cubic',
 #'     'sinc70', or 'sinc700'.
+#' @param skip.errors Sometimes, for some segments, Praat fails to create an Intensity
+#'     object. If skip.errors = TRUE, analysis those segments will be skipped, and corresponding
+#'     pitch values will be returned as "--undefined--". If skip.errors = FALSE, the error
+#'     message from Praat will be returned in the Error field, but no pitch measures will
+#'     be returned for any segments in the same recording.
 #' @return A script fragment which can be passed as the praat.script parameter of
 #'     \link{processWithPraat} 
 #' 
@@ -42,7 +47,7 @@
 #' }
 #' @keywords praat
 #' 
-praatScriptIntensity <- function(minimum.pitch = 100.0, time.step = 0.0, subtract.mean = TRUE, get.maximum = TRUE, sample.points = NULL, interpolation = 'cubic') {
+praatScriptIntensity <- function(minimum.pitch = 100.0, time.step = 0.0, subtract.mean = TRUE, get.maximum = TRUE, sample.points = NULL, interpolation = 'cubic', skip.errors = TRUE) {
     if (subtract.mean) {
         script <- "\nsubtractmean$ = \"yes\""
     } else {
@@ -51,12 +56,29 @@ praatScriptIntensity <- function(minimum.pitch = 100.0, time.step = 0.0, subtrac
     script <- paste( # ensure the sound sample is selected
         script, "\nselect Sound 'sampleName$'", sep="")
     
+    if (skip.errors) { ## use nocheck
+        script <- paste(
+            script,
+            "\n# nocheck to prevent the whole script from failing, and then check for object after",
+            "\nnocheck ", sep="")
+    } else {
+        script <- paste(script, "\n", sep="")
+    }
+    
     script <- paste(
-        script, "\nTo Intensity: ", minimum.pitch, ", ", time.step, ", subtractmean$", sep="")
+        script, "To Intensity: ", minimum.pitch, ", ", time.step, ", subtractmean$", sep="")
 
+    script <- paste(
+        script,
+        "\n# check that an Intensity object was created",
+        "\nobjectCreated = extractWord$(selected$(), \"\") = \"Intensity\"", sep="")
     if (get.maximum) {
         script <- paste(script,
-                        "\nmaxIntensity = Get maximum: targetStart, targetEnd, \"Parabolic\"",
+                        "\nif objectCreated",
+                        "\n  maxIntensity = Get maximum: targetStart, targetEnd, \"Parabolic\"",
+                        "\nelse",
+                        "\n  maxIntensity = 1/0", # --undefined--
+                        "\nendif",
                         "\nprint 'maxIntensity' 'newline$'", sep="")
     }
     
@@ -73,14 +95,22 @@ praatScriptIntensity <- function(minimum.pitch = 100.0, time.step = 0.0, subtrac
             script <- paste(script, "\npointoffset =",
                             " targetStart + ", point, " * targetDuration", sep="")
             varname = paste("intensity_time_", stringr::str_replace(point, "\\.","_"), sep="")
-            script <- paste(script, "\n", varname,
-                           " = Get value at time: pointoffset, \"",interpolation,"\"",
-                           sep="")
+            script <- paste(script,
+                            "\nif objectCreated",
+                            "\n  ", varname,
+                            " = Get value at time: pointoffset, \"",interpolation,"\"",
+                            "\nelse",
+                            "\n  ", varname, " = 1/0", # --undefined--
+                            "\nendif",
+                            sep="")
             script <- paste(script, "\nprint '", varname, ":0' 'newline$'", sep="")
         } ## next sample point
     }
     
     script <- paste( # remove intensity object
-        script, "\nRemove\n", sep="")
+        script,
+        "\nif objectCreated",
+        "\n  Remove",
+        "\nendif\n", sep="")
     return(script)
 }